requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
